TIP-843: Rename array converter specs to enrichment
